langgraph-react/pyproject.toml
[tool.poetry]
name = "langgraph-react"
version = "0.1.0"
description = ""
authors = ["adamscarlat <ascarlat@vividcloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.2.7"
langgraph = "^0.1.8"
langchain-openai = "^0.1.15"
langchainhub = "^0.1.20"
python-dotenv = "^1.0.1"
langchain-community = "^0.2.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
